Don't scan the whole bus to verify existence of i2c device
Instead of sending an empty write request to every possible address on
the but to see which devices respond, send it only to the one address
that we are checking. This is not only much faster and more robust
(there might be devices on the bus that respond badly to such requests
at the wrong speed) but also makes it much easier to debug things with a
logic analyzer, as it doesn't get flooded with all those requests.