Fix I2C init error on BeagleBone Black
This init method verifies that a device exists at the given address.

It was writing a zero byte value to the bus.  This triggered a
bug in the Linux kernel I2C bus OMAP2 driver for the BeagleBone.
The driver returns an invalid write error when msg->len is 0:
https://github.com/beagleboard/linux/blob/4.14/drivers/i2c/busses/i2c-omap.c#L665

The solution is to write the value 'x' instead of ''.

Refer to Adafruit_Blinka PR #42 for more information:
https://github.com/adafruit/Adafruit_Blinka/pull/42#issuecomment-431948703